move method to pathresolver
